UI 개선

+++ src/dashboard/components/DetailPanel/DetailPanel.js
@@ -13,27 +13,28 @@ class DetailPanel extends Component {
     userId: window.localStorage.getItem("user_id"),
     port: null,
     open: false,
-    del_trgt: null,
+    target: null,
+    action: null
   };
   
   // DOM 마운트 후
   componentDidMount() {
+    this.setState({
+      loadOfDatas : true
+    });
     this._getContainer();
   }
 
   // 컨테이너 조회
   _getContainer = async () => {
     const res = await axios.post("/api/search", {userId : window.localStorage.getItem("user_id")});
-    console.log(res);
     var arr = [];
 
     for (let element of res.data.container) {
       let info = await axios.get("/containers/" + element.container_id + "/json");
       arr.push(info.data.State.Running);
     }
-
-    console.log(arr);
-
+    console.log(res.data.container);
     this.setState({ container: res.data.container, container_status : arr, loadOfDatas: false});
   };
 
@@ -48,36 +49,26 @@ class DetailPanel extends Component {
   }
 
   // 컨테이너 삭제
-  handleDelete = async (e) => {
-    let data = this.state.container[this.state.del_trgt];
+  handleStopAndDelete = async (e) => {
+    let data = this.state.container[this.state.target];
     this.setState({ loadOfDatas: true, open: false});
 
     // 컨테이너 정지 후 제거
     try {
       await axios.post("/containers/" + data.container_id + "/stop");
-      await axios.delete("/containers/" + data.container_id);
+      
+      if(this.state.action === "delete"){
+        await axios.delete("/containers/" + data.container_id);
+      }
     } catch (e) {
       console.log(e);
     } finally {
-      await axios.post("/api/delete", data);
+      if(this.state.action === "delete"){
+        await axios.post("/api/delete", data);
+      }
       await this._getContainer();
     }
   };
-
-  // 컨테이너 정지
-  stopContainer = async (e) => {
-    let data = this.state.container[e.target.value];
-    
-    this.setState({ loadOfDatas: true});
-    
-    try {
-      await axios.post("/containers/" + data.container_id + "/stop");
-    } catch(e){
-      console.log(e);
-    } finally {
-      await this._getContainer();
-    }
-  }
 
   // 컨테이너 실행
   startContainer = async(e) => {
@@ -94,10 +85,10 @@ class DetailPanel extends Component {
         
           setTimeout(function () {
             this._getContainer();
-            window.open("http://es2.danawa.io:3333/" + window.localStorage.getItem("user_id") + "/" + item.port +"/?folder=/home/danawa/works/" +   (item.stack_cd === "java" ? "spring-boot-helloworld-master" : "node-js-sample-master") + (item.tmlt_dtl === 'non' ? "/clone_space" : ""), "_blank");
+            window.open("http://es2.danawa.io:3333/" + window.localStorage.getItem("user_id") + "/" + item.port +"/?folder=/home/danawa/works/" +   (item.stack_cd === "java" ? "spring-boot-helloworld-master" : "node-js-sample-master"), "_blank");
           }.bind(this), 8000)
         } else {
-          window.open("http://es2.danawa.io:3333/" + window.localStorage.getItem("user_id") + "/" + item.port +"/?folder=/home/danawa/works/" + (item.stack_cd === "java" ? "spring-boot-helloworld-master" : "node-js-sample-master") + (item.tmlt_dtl === 'non' ? "/clone_space" : ""), "_blank");
+          window.open("http://es2.danawa.io:3333/" + window.localStorage.getItem("user_id") + "/" + item.port +"/?folder=/home/danawa/works/" + (item.stack_cd === "java" ? "spring-boot-helloworld-master" : "node-js-sample-master"), "_blank");
         }
       } 
     } catch (e) {
@@ -107,7 +98,7 @@ class DetailPanel extends Component {
 
 
   handleCancel = () => this.setState({open: false, loadOfDatas: false });
-  show = (e) => { console.log(e.target.value); this.setState({ open: true, del_trgt : e.target.value}); } 
+  show = (e) => {this.setState({ open: true, target : e.target.value, action : e.target.name}); } 
 
   render() {
     var containerList = [];
@@ -115,13 +106,15 @@ class DetailPanel extends Component {
     if(this.state.container !== null) {
       containerList = this.state.container.map((item, i) => ( 
         <div
-            className="inner-content"
+            className="card_grid"
             key={i}
           >
             <Card className="container">
               <Card.Content header={item.container_nm} />
+
               <Card.Content
                 id="card"
+                meta= {"생성일시 : " + item.insert_dts.substring(0, 10) + " " + item.insert_dts.substring(11, 16)}
                 description={item.note_txt}
               />
               <Card.Content extra>
@@ -129,10 +122,15 @@ class DetailPanel extends Component {
                   언어
                   <Label.Detail>{item.stack_cd}</Label.Detail>
                 </Label>
-                <Label className="container-text-zone" color="violet">
+                <Label className="container-text-zone" color="green" style={this.state.container_status[i] === true ? {display:"block"} : {display:"none"}}>
                   상태
-                  <Label.Detail>{this.state.container_status[i] === true ? "Running" : "Stop"}</Label.Detail>
+                  <Label.Detail>Running</Label.Detail>
                 </Label>
+                <Label className="container-text-zone" color="red" style={this.state.container_status[i] === true ? {display:"none"} : {display:"block"}}>
+                  상태
+                  <Label.Detail>Stop</Label.Detail>
+                </Label>              
+
                 <Button
                   className="content-button"
                   content="▶ 터미널 실행"
@@ -140,11 +138,12 @@ class DetailPanel extends Component {
                   value={i}
                   onClick={this.startContainer}
                 ></Button>
-                <Button icon className="content-button-stop" color="grey" value={i} onClick={this.stopContainer}>정지</Button>
+                <Button icon className="content-button-stop" color="grey" name="stop" value={i} onClick={this.show}>정지</Button>
               </Card.Content>
               <Button
                 color="blue"
                 className="delete-button"
+                name="delete"
                 content="삭제"
                 size="mini"
                 value={i}
@@ -152,12 +151,12 @@ class DetailPanel extends Component {
               />
               <Confirm
                 open={this.state.open}
-                header="컨테이너 삭제"
-                content="컨테이너를 삭제하시겠습니까?"
+                header={this.state.action == "delete" ? "컨테이너 삭제" : "컨테이너 정지"}
+                content={this.state.action == "delete" ? "컨테이너를 삭제하시겠습니까?" : "컨테이너를 정지하시겠습니까?"}
                 cancelButton="취소"
-                confirmButton="삭제"
+                confirmButton= {this.state.action == "delete" ? "삭제" : "정지"}
                 onCancel={this.handleCancel}
-                onConfirm={this.handleDelete}
+                onConfirm= {this.handleStopAndDelete}
               />
             </Card>
           </div>
@@ -166,15 +165,16 @@ class DetailPanel extends Component {
 
     return (
       <div className="DetailPanel">
-        <h3>컨테이너</h3>
-        
-        <Link to={{pathname : '/newContainer', state : this.state}}>
-          <Button
-            className="Navigate-right-button"
-            color="grey"
-            content="+ 새 컨테이너"
-          />
-        </Link>
+        <div className="container-header">
+          <h3>Container
+          <Link to={{pathname : '/newContainer', state : this.state}}>
+            <Button
+              className="Navigate-right-button"
+              color="grey"
+            ><Icon name="add"/> NEW</Button>
+          </Link>
+          </h3>
+        </div>
 
         <Dimmer active={this.state.loadOfDatas}>
           <Loader></Loader>
@@ -187,7 +187,7 @@ class DetailPanel extends Component {
             <h4 className="content-text">
               Create More..
             </h4>
-            <img src={img} width="300" height="300" alt="새 컨테이너" />
+            <img src={img} width="290" height="300" alt="새 컨테이너" />
           </div>
         </Link>
       </div>
